Replaced is_active() with wait_for_finish().
We need this function due to problems with file sharing under
Windows. The main thread cannot delete files open in the child thread
until that child thread dies. Monitors and flags will not help as we need
to know when the thread actuall dies, and not when it's about to die and
is telling everybody about it.