Modified GString::format(), and added GNativeString::format(), and GUTF8String::format().
The format() method now accepts messages in the windows FormatMessage
syntax.  The syntax %n!...! will be converted to %n$... under Unix.
This allows arguments to be rearranged in the format string.

This allows arguments to be rearranged in the format string.

The string "EOF" is now refered to from the memory address
ByteStream::EndOfFile.