Updated the configuration scripts to correctly find libdl.a.
Updated the configuration scripts to use the -imacros gcc option
with a config.h file.

Fixed to DjVuOCR, DjVuFileList, djvurtk, to build a working DjVuOCR
codec, and various bugs related to using such a codec.

Fixed deadline and recursive loop problems in GString and GURL.

Corrected the logic for following the search path in DjVuMessage.cpp.